Change a bunch a size_t into int32_t or int64_t since it causes ambiguous type specialization when cast to GMP integers.

+++ linbox/algorithms/block-wiedemann.h
@@ -79,12 +79,13 @@ namespace LinBox
 			m = A.rowdim();
 			n = A.coldim();
 
-			size_t p,q;
+			uint32_t p,q;
+                        // CP : converting to GMP int to get the bitsize is unsane ! Should be replaced by a tablelookup
 			integer tmp;
-			tmp = m;
+			tmp = uint32_t(m);
 			p = tmp.bitsize()-1;
 			//p=sqrt(tmp);
-			tmp = n;
+			tmp = uint32_t(n);
 			q = tmp.bitsize()-1;
 			//q=sqrt(tmp);
 			//std::cout<<"row block: "<<p<<std::endl;
